// js/src/types/generated/cdp.ts
export const REQUEST = "request" as const;
export const RESPONSE = "response" as const;
export const EVENT = "event" as const;

export { zod, commands, events };
export type CdpRuntime = {
  readonly types: typeof runtimeTypes;
  readonly commands: typeof commands;
  readonly events: typeof events;
  readonly REQUEST: typeof REQUEST;
  readonly RESPONSE: typeof RESPONSE;
  readonly EVENT: typeof EVENT;
};
export const cdp: CdpRuntime = { types: runtimeTypes, commands, events, REQUEST, RESPONSE, EVENT };
export const types: typeof runtimeTypes = cdp.types;

export namespace cdp {
  export namespace types {
    export namespace ts {
// ...
      export namespace Target {
        export type TargetID = z.infer<typeof zod.Target.TargetID>;
        export type SessionID = z.infer<typeof zod.Target.SessionID>;
        export type TargetInfo = z.infer<typeof zod.Target.TargetInfo>;
        export type FilterEntry = z.infer<typeof zod.Target.FilterEntry>;
        export type TargetFilter = z.infer<typeof zod.Target.TargetFilter>;
        export type RemoteLocation = z.infer<typeof zod.Target.RemoteLocation>;
        export type WindowState = z.infer<typeof zod.Target.WindowState>;
        export type ActivateTargetParams = z.infer<typeof zod.Target.ActivateTargetParams>;
        export type ActivateTargetResult = z.infer<typeof zod.Target.ActivateTargetResult>;
        export type AttachToTargetParams = z.infer<typeof zod.Target.AttachToTargetParams>;
        export type AttachToTargetResult = z.infer<typeof zod.Target.AttachToTargetResult>;
        export type AttachToBrowserTargetParams = z.infer<typeof zod.Target.AttachToBrowserTargetParams>;
        export type AttachToBrowserTargetResult = z.infer<typeof zod.Target.AttachToBrowserTargetResult>;
        export type CloseTargetParams = z.infer<typeof zod.Target.CloseTargetParams>;
        export type CloseTargetResult = z.infer<typeof zod.Target.CloseTargetResult>;
        export type ExposeDevToolsProtocolParams = z.infer<typeof zod.Target.ExposeDevToolsProtocolParams>;
        export type ExposeDevToolsProtocolResult = z.infer<typeof zod.Target.ExposeDevToolsProtocolResult>;
        export type CreateBrowserContextParams = z.infer<typeof zod.Target.CreateBrowserContextParams>;
        export type CreateBrowserContextResult = z.infer<typeof zod.Target.CreateBrowserContextResult>;
        export type GetBrowserContextsParams = z.infer<typeof zod.Target.GetBrowserContextsParams>;
        export type GetBrowserContextsResult = z.infer<typeof zod.Target.GetBrowserContextsResult>;
        export type CreateTargetParams = z.infer<typeof zod.Target.CreateTargetParams>;
        export type CreateTargetResult = z.infer<typeof zod.Target.CreateTargetResult>;
        export type DetachFromTargetParams = z.infer<typeof zod.Target.DetachFromTargetParams>;
        export type DetachFromTargetResult = z.infer<typeof zod.Target.DetachFromTargetResult>;
        export type DisposeBrowserContextParams = z.infer<typeof zod.Target.DisposeBrowserContextParams>;
        export type DisposeBrowserContextResult = z.infer<typeof zod.Target.DisposeBrowserContextResult>;
        export type GetTargetInfoParams = z.infer<typeof zod.Target.GetTargetInfoParams>;
        export type GetTargetInfoResult = z.infer<typeof zod.Target.GetTargetInfoResult>;
        export type GetTargetsParams = z.infer<typeof zod.Target.GetTargetsParams>;
        export type GetTargetsResult = z.infer<typeof zod.Target.GetTargetsResult>;
        export type SendMessageToTargetParams = z.infer<typeof zod.Target.SendMessageToTargetParams>;
        export type SendMessageToTargetResult = z.infer<typeof zod.Target.SendMessageToTargetResult>;
        export type SetAutoAttachParams = z.infer<typeof zod.Target.SetAutoAttachParams>;
        export type SetAutoAttachResult = z.infer<typeof zod.Target.SetAutoAttachResult>;
        export type AutoAttachRelatedParams = z.infer<typeof zod.Target.AutoAttachRelatedParams>;
        export type AutoAttachRelatedResult = z.infer<typeof zod.Target.AutoAttachRelatedResult>;
        export type SetDiscoverTargetsParams = z.infer<typeof zod.Target.SetDiscoverTargetsParams>;
        export type SetDiscoverTargetsResult = z.infer<typeof zod.Target.SetDiscoverTargetsResult>;
        export type SetRemoteLocationsParams = z.infer<typeof zod.Target.SetRemoteLocationsParams>;
        export type SetRemoteLocationsResult = z.infer<typeof zod.Target.SetRemoteLocationsResult>;
        export type GetDevToolsTargetParams = z.infer<typeof zod.Target.GetDevToolsTargetParams>;
        export type GetDevToolsTargetResult = z.infer<typeof zod.Target.GetDevToolsTargetResult>;
        export type OpenDevToolsParams = z.infer<typeof zod.Target.OpenDevToolsParams>;
        export type OpenDevToolsResult = z.infer<typeof zod.Target.OpenDevToolsResult>;
        export type AttachedToTargetEvent = z.infer<typeof zod.Target.AttachedToTargetEvent>;
        export type DetachedFromTargetEvent = z.infer<typeof zod.Target.DetachedFromTargetEvent>;
        export type ReceivedMessageFromTargetEvent = z.infer<typeof zod.Target.ReceivedMessageFromTargetEvent>;
        export type TargetCreatedEvent = z.infer<typeof zod.Target.TargetCreatedEvent>;
        export type TargetDestroyedEvent = z.infer<typeof zod.Target.TargetDestroyedEvent>;
        export type TargetCrashedEvent = z.infer<typeof zod.Target.TargetCrashedEvent>;
        export type TargetInfoChangedEvent = z.infer<typeof zod.Target.TargetInfoChangedEvent>;
      }
// ...
    }
  }
}