// js/src/types/generated/zod/Target.ts
export const TargetID = withCdpMeta(z.string(), "Target.TargetID", "type");
export const SessionID = withCdpMeta(z.string(), "Target.SessionID", "type");
export const TargetInfo = withCdpMeta(z.object({ "targetId": z.lazy(() => TargetID), "type": z.string(), "title": z.string(), "url": z.string(), "attached": z.boolean(), "parentId": z.lazy(() => TargetID).optional(), "openerId": z.lazy(() => TargetID).optional(), "canAccessOpener": z.boolean(), "openerFrameId": z.lazy(() => Page.FrameId).optional(), "parentFrameId": z.lazy(() => Page.FrameId).optional(), "browserContextId": z.lazy(() => Browser.BrowserContextID).optional(), "subtype": z.string().optional() }).passthrough(), "Target.TargetInfo", "type");
export const FilterEntry = withCdpMeta(z.object({ "exclude": z.boolean().optional(), "type": z.string().optional() }).passthrough(), "Target.FilterEntry", "type");
export const TargetFilter = withCdpMeta(z.array(z.lazy(() => FilterEntry)), "Target.TargetFilter", "type");
// ...
export const GetTargetsParams = withCdpMeta(z.object({ "filter": z.lazy(() => TargetFilter).optional() }).passthrough(), "Target.getTargets.params", "commandParams", { method: "Target.getTargets" });
export const GetTargetsResult = withCdpMeta(z.object({ "targetInfos": z.array(z.lazy(() => TargetInfo)) }).passthrough(), "Target.getTargets.result", "commandResult", { method: "Target.getTargets" });
export const GetTargetsCommand = withCdpCommand("Target.getTargets", GetTargetsParams, GetTargetsResult);
// ...
export const SetAutoAttachParams = withCdpMeta(z.object({ "autoAttach": z.boolean(), "waitForDebuggerOnStart": z.boolean(), "flatten": z.boolean().optional(), "filter": z.lazy(() => TargetFilter).optional() }).passthrough(), "Target.setAutoAttach.params", "commandParams", { method: "Target.setAutoAttach" });
export const SetAutoAttachResult = withCdpMeta(z.object({  }).passthrough(), "Target.setAutoAttach.result", "commandResult", { method: "Target.setAutoAttach" });
export const SetAutoAttachCommand = withCdpCommand("Target.setAutoAttach", SetAutoAttachParams, SetAutoAttachResult);
// ...
export const AttachedToTargetEvent = withCdpMeta(z.object({ "sessionId": z.lazy(() => SessionID), "targetInfo": z.lazy(() => TargetInfo), "waitingForDebugger": z.boolean() }).passthrough(), "Target.attachedToTarget", "event", { phase: "event" });
export const DetachedFromTargetEvent = withCdpMeta(z.object({ "sessionId": z.lazy(() => SessionID), "targetId": z.lazy(() => TargetID).optional() }).passthrough(), "Target.detachedFromTarget", "event", { phase: "event" });